added dynamic neuron color based on its bias

+++ NeuralNetworkFX/src/application/InputNeuron.java
@@ -15,8 +15,8 @@ public class InputNeuron implements Neuron {
 	
 	@SuppressWarnings("null")	// warning suppressed because fully unfixable
 	@Override
-	public double getBias() {
-		return  (Double) null;
+	public Double getBias() {
+		return  null;
 	}
 	@Override
 	public void setBias(double bias) {
